Improvements in controllers, services and repositories. Additional methods added in services. IsDeleted property added for Product Category with migrations. Enums for order statuses added. Other code improvements

+++ EcommerceStore.Application/Models/InputModels/ProductDetailsForOrderInputModel.cs
@@ -8,7 +8,7 @@ namespace EcommerceStore.Application.Models.InputModels
         [Range(1, int.MaxValue, ErrorMessage = "Product id must be an integer starting from 1")]
         public int ProductId { get; set; }
         [Required]
-        [Range(0.1, double.MaxValue, ErrorMessage = "Product amount must be positive integer")]
+        [Range(0.1, int.MaxValue, ErrorMessage = "Product amount must be positive integer")]
         public int ProductAmount { get; set; }
     }
 }
